Addition fast task

+++ src/models/fast_task.go
@@ -4,7 +4,7 @@ import "time"
 
 type FastTask struct {
 	Id         int           `json:"id"`
-	AssigneeId int64         `json:"assignee_id"`
+	AssigneeId int           `json:"assignee_id"`
 	TaskName   string        `json:"task_name"`
 	Interval   time.Duration `json:"interval"`
 }
